feat(megatron): migrate dpo to native training entry
